<update> inventory combine quantity

+++ GameLogic/src/main/java/nro/model/item/ItemTemplate.java
@@ -3,7 +3,7 @@ package nro.model.item;
 import java.util.List;
 
 public record ItemTemplate(short id, byte type, byte gender, String name, String description, byte level, short iconID,
-                           short part, byte maxQuantity, int strRequire, short head, short body, short leg, List<ItemOption> options) {
+                           short part, int maxQuantity, int strRequire, short head, short body, short leg, List<ItemOption> options) {
 
     public record ArrHead2Frames(int id, List<Integer> frames) {
     }
